requirements.txt
python-telegram-bot==20.3
openai==1.2.3